\name{print.simtest.ratio}
\alias{print.simtest.ratio}

\title{ Print out the results of simtest.ratio  }
\description{
  A short print out of the results of simtest.ratio
}
\usage{
\method{print}{simtest.ratio}(x, digits = 4, ...)
}

\arguments{
  \item{x}{ An object of class "simtest.ratio" as obtained by calling simtest.ratio }
  \item{digits}{ digits for rounding of the results }
  \item{\dots}{arguments to be passed to print}
}
\value{
A print out, containing the margins, estimates, teststatistics, and p.values computed by simtest.ratio. 
}


\keyword{ print }