\name{summary.simtest.ratio}
\alias{summary.simtest.ratio}
\title{ Summary function for simtest.ratio }
\description{
  A detailed print out of the results of simtest.ratio}


\usage{
\method{summary}{simtest.ratio}(object, digits = 4, ...)
}

\arguments{
  \item{object}{ An object of class "simtest.ratio" as obtained by calling simtest.ratio }
  \item{digits}{ digits for rounding of the results }
  \item{\dots}{arguments to be passed to print}

}
\value{
A print out, containing the numerator and denominator contrast matrices, the correlation under the null-hypothesis, margins, estimates, teststatistics, and p.values computed by simtest.ratio. 
}


\keyword{ print }



